[AMP-2553] Fix tests
