+++ site/components/src/test/java/uk/nhs/digital/common/components/catalogue/ApiCatalogueComponentTest.java
@@ -42,10 +42,7 @@ import uk.nhs.digital.website.beans.ComponentList;
 import uk.nhs.digital.website.beans.Externallink;
 import uk.nhs.digital.website.beans.Internallink;
 
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @RunWith(PowerMockRunner.class)
@@ -64,6 +61,7 @@ public class ApiCatalogueComponentTest extends MockitoSessionTestBase {
     @Mock private CatalogueRepository catalogueRepository;
     @Mock private Filters expectedFiltersFromFactory;
     @Mock private FiltersFactory filtersFactory;
+    @Mock private FacetNavHelper facetNavHelper;
 
     private HstContainerURL hstContainerUrl;
     private MockHstRequest request;
@@ -95,6 +93,10 @@ public class ApiCatalogueComponentTest extends MockitoSessionTestBase {
         irrelevantResponse = new MockHstResponse();
 
         apiCatalogueComponent = new ApiCatalogueComponent();
+
+        apiCatalogueComponent.facetNavHelper = facetNavHelper;
+        when(facetNavHelper.getAllTags()).thenReturn(allTags());
+        when(facetNavHelper.getAllTagsForLink(any())).thenReturn(new ArrayList<String>(allTags()));
     }
 
     @Test
@@ -102,6 +104,7 @@ public class ApiCatalogueComponentTest extends MockitoSessionTestBase {
 
         // when
         when(expectedFiltersFromFactory.initialisedWith(any(), any())).thenReturn(expectedFiltersFromFactory);
+
         apiCatalogueComponent.doBeforeRender(request, irrelevantResponse);
 
         // then
@@ -283,6 +286,10 @@ public class ApiCatalogueComponentTest extends MockitoSessionTestBase {
                 actualResults.size(),
                 is(7)
         );
+    }
+
+    private Set<String> allTags() {
+        return new HashSet<String>(Arrays.asList("hl7-v3", "dental-health", "fhir", "inpatient", "mental-health", "hospital", "deprecated-api", "retired-api"));
     }
 
     @SuppressWarnings({"unchecked", "rawtypes"})
